+++ src/shared/components/CheckBox/index.tsx
